chore: refactor remove nesting
